requirements.txt
strands-agents>=0.1.0
strands-agents-tools>=0.1.0
boto3>=1.35.0
python-dotenv>=1.0.0
requests>=2.31.0
